requirements.txt
DearPyGui-DragAndDrop
dearpygui
numpy
pillow
# pillow-simd  # Does not support python > 3.10
annotated-types
tqdm
darkdetect
win32mica
